requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kafka-python==2.0.2
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
pyyaml==6.0.1
sqlalchemy==2.0.23
python-dateutil==2.8.2
boto3==1.34.51
pyarrow==15.0.0
redis==5.0.1
requests==2.31.0

